#4402 remove extra rule status file
